requirements.txt
boto3==1.37.28
fastapi==0.105.0
uvicorn==0.24.0
gunicorn==21.2.0
instagrapi==2.1.3
python-jwt==4.0.0
supabase==1.2.0
apscheduler==3.10.4
python-multipart==0.0.6
python-dateutil==2.8.2
pydantic==2.10.1
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
requests==2.31.0
PyJWT==2.8.0
Pillow>=8.1.1
moviepy==1.0.3
cryptography>=41.0.5
python-dotenv==1.0.0
facebook-sdk==3.1.0
psycopg2-binary==2.9.9
alembic==1.12.1
